move SearchQuery and SortComparator into 'search' package
